Refactored streamfunction folder -> external  reason: the streamfunction is a subset of external mode solvers along with pressure solver
