+++ veros/core/external/__init__.py
@@ -1,0 +1,2 @@
+from veros.core.external.streamfunction_init import streamfunction_init  # noqa: F401
+from veros.core.external.solve_stream import solve_streamfunction  # noqa: F401
