+++ veros/core/external/solve_pressure.py
@@ -15,7 +15,7 @@ from veros.variables import allocate
 from veros.core import utilities as mainutils
 from veros.core.operators import update, update_add, at, for_loop
 from veros.core.operators import numpy as npx
-from veros.core.streamfunction.pressure_solvers import get_linear_solver
+from veros.core.external.pressure_solvers import get_linear_solver
 
 
 @veros_routine
