+++ veros/core/external/solvers/base.py
