--- veros/core/streamfunction/__init__.py
@@ -1,2 +1,0 @@
-from veros.core.streamfunction.streamfunction_init import streamfunction_init  # noqa: F401
-from veros.core.streamfunction.solve_stream import solve_streamfunction  # noqa: F401
